Extract out Spatial submodule.
