+++ test/runtests.jl
@@ -17,7 +17,7 @@ create_autodiff(x, dx) = [ForwardDiff.Dual(x[i], dx[i]) for i in 1 : length(x)]
 @inline Base.rem(x::ForwardDiff.Dual, n::Real) = ForwardDiff.Dual(rem(ForwardDiff.value(x), n), ForwardDiff.partials(x))
 
 include("test_graph.jl")
-include("test_util.jl")
+include("test_custom_collections.jl")
 include("test_frames.jl")
 include("test_spatial.jl")
 include("test_contact.jl")
